refactor(e2e): move the browser-launch helper out of conftest
Importing `e2e.dash.conftest` from a test in the main suite executed that
module's import-time side effect -- it sets DJANGO_ALLOW_ASYNC_UNSAFE for the
sync-Playwright + sync-Django combination, which has no business leaking into
other lanes. Disabling Django's async-unsafe guard suite-wide plausibly widens
the window for the known leaked-sqlite-connection unraisable flake, and this
branch drew five failing shards against one to three elsewhere.

The helper moves to a side-effect-free `e2e/dash/browser_launch.py`; conftest
imports it, and the test imports the pure module. Verified: importing the new
module leaves DJANGO_ALLOW_ASYNC_UNSAFE unset.
